authentication done here

+++ api/app.js
@@ -8,37 +8,50 @@ const { List, Task, User } = require('./db/models');
 const jwt = require('jsonwebtoken');
 
 const app = express()
-    .use(cors())
+    .use((req, res, next) => {
+        res.header("Access-Control-Allow-Origin", "*"); // update to match the domain you will make the request from
+        res.header("Access-Control-Allow-Methods", "GET, POST, HEAD, OPTIONS, PUT, PATCH, DELETE");
+        res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept, x-access-token, x-refresh-token, _id");
+        res.header('Access-Control-Expose-Headers', 'x-access-token, x-refresh-token');
+        next();
+    })
     .use(express.json())
     .use(express.urlencoded({ extended: true }));
 
 mongoose.connect(
-    'mongodb://localhost:27017/TasManager', 
+    'mongodb://localhost:27017/TasManager',
     { useNewUrlParser: true, useCreateIndex: true, useFindAndModify: false },
     (error, response) => {
-        if(response) {
+        if (response) {
             console.log("Connected to MongDB successfully");
         }
 
-        if(error) {
+        if (error) {
             console.log("Error while attempting to connect to MongoDB");
             console.log(error);
         }
     }
 )
-app.use(function (req, res, next) {
-    res.header("Access-Control-Allow-Origin", "*"); // update to match the domain you will make the request from
-    res.header("Access-Control-Allow-Methods", "GET, POST, HEAD, OPTIONS, PUT, PATCH, DELETE");
-    res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept, x-access-token, x-refresh-token, _id");
-    res.header(
-        'Access-Control-Expose-Headers',
-        'x-access-token, x-refresh-token'
-    );
-    next();
-});
+
+// check whether the request has a valid JWT access token
+let authenticate = (req, res, next) => {
+    let token = req.header('x-access-token');
+    // verify the JWT
+    jwt.verify(token, User.getJWTSecret(), (err, decoded) => {
+        if (err) {
+            // there was an error
+            // jwt is invalid - * DO NOT AUTHENTICATE *
+            res.status(401).send(err);
+        } else {
+            // jwt is valid
+            req.user_id = decoded._id;
+            next();
+        }
+    });
+}
 
 // Verify Refresh Token Middleware (which will be verifying the session)
-let verifySession =(req, res, next) => {
+let verifySession = (req, res, next) => {
     // grab the refresh token from the request header
     let refreshToken = req.header('x-refresh-token');
     // grab the _id from the request header
@@ -91,10 +104,14 @@ let verifySession =(req, res, next) => {
  */
 
 app.get('/', (req, res) => { res.send(" Hello ") })
-app.get('/lists', (req, res) => {
-    // we want to return an array of all the lists in the database
-    List.find({}).then((lists) => {
+app.get('/lists', authenticate, (req, res) => {
+    // we want to return an array of all the lists in the database that belong to authenticated user
+    List.find({
+        _userId: req.user_id
+    }).then((lists) => {
         res.send(lists);
+    }).catch((e) => {
+        res.send(e);
     });
 })
 
@@ -102,12 +119,13 @@ app.get('/lists', (req, res) => {
  * POST /lists
  * Purpose: Create a list
  */
-app.post('/lists', (req, res) => {
+app.post('/lists', authenticate, (req, res) => {
     // we want to create a new list and return the new list document back to the user(which include the id)
     // the list information(fields) will be passed in via the JSON request body
     let title = req.body.title;
     let newList = new List({
-        title
+        title,
+        _userId: req.user_id
     });
     newList.save().then((listDoc) => {
         res.send(listDoc);
@@ -118,9 +136,9 @@ app.post('/lists', (req, res) => {
  * PATCH /lists/:id
  * Purpose: Update a specified list 
  */
-app.patch('/lists/:id', (req, res) => {
+app.patch('/lists/:id', authenticate, (req, res) => {
     // we want to update the specified list(list document with id in the URL) with the new values specified in the JSON body of the request
-    List.findOneAndUpdate({ _id: req.params.id }, {
+    List.findOneAndUpdate({ _id: req.params.id, _userId: req.user_id }, {
         $set: req.body
     }).then(() => {
         res.sendStatus(200);
@@ -131,12 +149,15 @@ app.patch('/lists/:id', (req, res) => {
  * DELETE /lists/:id
  * Purpose: Delete a list
  */
-app.delete('/lists/:id', (req, res) => {
+app.delete('/lists/:id', authenticate, (req, res) => {
     // we want to delete the specified list(document with id in the URL)
     List.findOneAndRemove({
-        _id: req.params.id
+        _id: req.params.id,
+        _userId: req.user_id
     }).then((removedListDoc) => {
         res.send(removedListDoc);
+        // delete all the tasks that are in the deleted list
+        deleteTasksFromList(removedListDoc._id);
     })
 });
 
@@ -144,9 +165,9 @@ app.delete('/lists/:id', (req, res) => {
  * GET /lists/:listId/tasks
  * Purpose: Get all tasks in a specific list
  */
-app.get('/lists/:listId/tasks', (req, res) => {
+app.get('/lists/:listId/tasks', authenticate, (req, res) => {
     // We want to return all tasks that belong to a specific list (specified by listId)
-    if(req.params.listId != undefined) {
+    if (req.params.listId != undefined) {
         Task.find({
             _listId: req.params.listId
         }).then((tasks) => {
@@ -159,14 +180,31 @@ app.get('/lists/:listId/tasks', (req, res) => {
  * POST /lists/:listId/tasks
  * Purpose: Create a new task in a specific list
  */
-app.post('/lists/:listId/tasks', (req, res) => {
+app.post('/lists/:listId/tasks', authenticate, (req, res) => {
     // We want to create a new task in a list specified by listId
-    let newTask = new Task({
-        title: req.body.title,
-        _listId: req.params.listId
-    });
-    newTask.save().then((newTaskDoc) => {
-        res.send(newTaskDoc);
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            // list object with the specified conditions was found
+            // therefore the currently authenticated user can create new tasks
+            return true;
+        }
+        // else - the list object is undefined
+        return false;
+    }).then((canCreateTask) => {
+        if (canCreateTask) {
+            let newTask = new Task({
+                title: req.body.title,
+                _listId: req.params.listId
+            });
+            newTask.save().then((newTaskDoc) => {
+                res.send(newTaskDoc);
+            })
+        } else {
+            res.sendStatus(404);
+        }
     })
 })
 
@@ -174,16 +212,34 @@ app.post('/lists/:listId/tasks', (req, res) => {
 * PATCH /lists/:listId/tasks/:taskId
 * Purpose: Update an existing task
 */
-app.patch('/lists/:listId/tasks/:taskId', (req, res) => {
+app.patch('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
     // We want to update an existing task (specified by taskId)
-    Task.findOneAndUpdate({
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }, {
-        $set: req.body
-    }
-    ).then(() => {
-        res.send({message: 'Updated Successfully'});
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            // list object with the specified conditions was found
+            // therefore the currently authenticated user can make updates to tasks within this list
+            return true;
+        }
+        // else - the list object is undefined
+        return false;
+    }).then((canUpdateTasks) => {
+        if (canUpdateTasks) {
+            // the currently authenticated user can update tasks
+            Task.findOneAndUpdate({
+                _id: req.params.taskId,
+                _listId: req.params.listId
+            }, {
+                $set: req.body
+            }
+            ).then(() => {
+                res.send({ message: 'Updated successfully.' })
+            })
+        } else {
+            res.sendStatus(404);
+        }
     })
 });
 
@@ -191,13 +247,30 @@ app.patch('/lists/:listId/tasks/:taskId', (req, res) => {
  * DELETE /lists/:listId/tasks/:taskId
  * Purpose: Delete a task
  */
-app.delete('/lists/:listId/tasks/:taskId', (req, res) => {
-    Task.findOneAndRemove({
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }).then((removedTaskDoc) => {
-        res.send(removedTaskDoc);
-    })
+app.delete('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            // list object with the specified conditions was found
+            // therefore the currently authenticated user can make updates to tasks within this list
+            return true;
+        }
+        // else - the list object is undefined
+        return false;
+    }).then((canDeleteTasks) => {
+        if (canDeleteTasks) {
+            Task.findOneAndRemove({
+                _id: req.params.taskId,
+                _listId: req.params.listId
+            }).then((removedTaskDoc) => {
+                res.send(removedTaskDoc);
+            })
+        } else {
+            res.sendStatus(404);
+        }
+    });
 });
 
 /* USER ROUTES */
@@ -206,7 +279,7 @@ app.delete('/lists/:listId/tasks/:taskId', (req, res) => {
  * POST /users
  * Purpose: Sign up
  */
- app.post('/users', (req, res) => {
+app.post('/users', (req, res) => {
     // User sign up
     let body = req.body;
     let newUser = new User(body);
@@ -234,7 +307,7 @@ app.delete('/lists/:listId/tasks/:taskId', (req, res) => {
  * POST /users/login
  * Purpose: Login
  */
- app.post('/users/login', (req, res) => {
+app.post('/users/login', (req, res) => {
     let email = req.body.email;
     let password = req.body.password;
     User.findByCredentials(email, password).then((user) => {
@@ -261,7 +334,7 @@ app.delete('/lists/:listId/tasks/:taskId', (req, res) => {
  * GET /users/me/access-token
  * Purpose: generates and returns an access token
  */
- app.get('/users/me/access-token',verifySession, (req, res) => {
+app.get('/users/me/access-token', verifySession, (req, res) => {
     // we know that the user/caller is authenticated and we have the user_id and user object available to us
     req.userObject.generateAccessAuthToken().then((accessToken) => {
         res.header('x-access-token', accessToken).send({ accessToken });
@@ -269,6 +342,15 @@ app.delete('/lists/:listId/tasks/:taskId', (req, res) => {
         res.status(400).send(e);
     });
 })
+
+/* HELPER METHODS */
+let deleteTasksFromList = (_listId) => {
+    Task.deleteMany({
+        _listId
+    }).then(() => {
+        console.log("Tasks from " + _listId + " were deleted!");
+    })
+}
 
 app.listen(3000, () => {
     console.log("Server is listening on port 3000");
